refactor(report): extract report filters and excel export into modules
