+++ apps/frontend/src/application/hooks/use-export.hook.ts
@@ -1,6 +1,9 @@
-import { exportAsJsonFile } from "#utils/export.ts";
-import type { ExtendedProject } from "#api/types/project.types.ts";
+import { exportAsExcelFile, exportAsJsonFile } from "#utils/export.ts";
+import type { ExtendedProject, ProjectReport } from "#api/types/project.types.ts";
 import { ProjectsActions } from "#application/actions/projects.actions.ts";
+import { buildReportExcelTabs } from "#utils/report-excel.ts";
+import { useAlert } from "#application/hooks/use-alert.hook.ts";
+import { useTranslation } from "react-i18next";
 import { useAppDispatch } from "./use-app-redux.hook.ts";
 
 /**
@@ -18,4 +21,35 @@ export const useProjectExport = () => {
     };
 
     return { exportProject };
+};
+
+/**
+ * Exports the full project report (assets, threats, measures, measure impacts) as an Excel file.
+ */
+export const useReportExcelExport = () => {
+    const { t } = useTranslation();
+    const { showErrorMessage } = useAlert();
+
+    const exportReportAsExcel = (
+        project: { name: string; confidentialityLevel: unknown },
+        reportData: ProjectReport | null
+    ) => {
+        if (!reportData) {
+            return;
+        }
+        const fileName =
+            Date.now() +
+            "_" +
+            project.name +
+            "-" +
+            String(project.confidentialityLevel ?? "").toUpperCase() +
+            "_export.xlsx";
+
+        exportAsExcelFile(buildReportExcelTabs(reportData), fileName).catch((error) => {
+            console.error("Excel export failed", error);
+            showErrorMessage({ message: t("errorMessages.excelExportFailed") });
+        });
+    };
+
+    return { exportReportAsExcel };
 };
